staticed a bunch of functions to GestureMovementHelpers.py
