+++ data_analysis/GestureClusterer.py
@@ -1,32 +1,14 @@
-
 #!/usr/bin/env pythons
-print("importing libs for GestureClusterer")
 import json
 import os
 import numpy as np
-import operator
 import random
 import time
 from tqdm import tqdm
 from sklearn.neighbors.nearest_centroid import NearestCentroid
-import matplotlib.pyplot as plt
-import statistics
-
+from GestureMovementHelpers import *
 from common_helpers import *
 
-
-BASE_KEYPOINT = [0]
-# DO NOT TRUST THESE
-RIGHT_BODY_KEYPOINTS = [1, 2, 3, 28]
-LEFT_BODY_KEYPOINTS = [4, 5, 6, 7]
-RIGHT_WRIST_KEYPOINT = 3
-LEFT_WRIST_KEYPOINT = 6
-#LEFT_HAND_KEYPOINTS = lambda x: [7] + [8 + (x * 4) + j for j in range(4)]  THESE ARE NOT RIGHT
-#RIGHT_HAND_KEYPOINTS = lambda x: [28] + [29 + (x * 4) + j for j in range(4)]   THESE ARE NOT RIGHT
-ALL_RIGHT_HAND_KEYPOINTS = list(range(31, 52))
-ALL_LEFT_HAND_KEYPOINTS = list(range(10, 31))
-BODY_KEYPOINTS = RIGHT_BODY_KEYPOINTS + LEFT_BODY_KEYPOINTS
-DIRECTION_ANGLE_SWITCH = 110  # arbitrary measure of degrees to constitute hands switching direction ¯\_(ツ)_/¯
 
 # semantics -- wn / tf
 # rhetorical
@@ -35,90 +17,37 @@ DIRECTION_ANGLE_SWITCH = 110  # arbitrary measure of degrees to constitute hands
 # prosody
 # motion dynamics
 
+# Now since we're only working within a single gesture, don't need to worry about
+# Mismatched timings anymore.
 
-# do difference between t1 and t2 for same frame,
-# and also difference between p1, p2, p3 etc at same frame
-# maybe get r_arm_distance_spread,      l_arm_distance_spread,     r_hand_distance_spread, l_hand_distance_spread
-#           [p1_t1 - p2_t1 - p3_t1...], [p1_t1 - p2_t1 - p3_t1...]
-#  ds_1 =               [t1                       t1...]
-#           [p1_t2 - p2_t2 - p3_t2...], [p1_t2 - p2_t2 - p3_t2...]
-#  ds_2 =               [t2                       t2...]
-# as represending [p1, p2, p3] etc at t1.
-# meanwhile each individual point also has a difference
-# [p0-p1], [p1-p2]
-#    t1      t2
-# p0-1_1 = [p0 - p1]
-# p0-1_2 = [p1 - p2]
-# so for each t you have
-# [ds_1, p0-1_1, p1-1_1, p2-1_1, p3-1_1...]
-# need to figure out how to compare subsequences of gesture.
-
-## the above is wayyyyy too complicated for now.
-## just get a gesture of high-level features, I think
-# All of the following refer to if it happens EVER in the gesture
-
-
-
-# [
-#   max_vert_accel,
-#   max_horiz_accel,
-#   l_palm_vert,
-#   l_palm_horiz,
-#   r_palm_vert,
-#   r_palm_horiz,
-#   x_oscillate,    # does it go up and down a lot?
-#   y_oscillate,
-#   hands_together,
-#   hands_separate,
-#   wrists_up,
-#   wrists_down,
-#   wrists_outward,
-#   wrists_inward,
-#   wrists_sweep,
-#   wrist_arc,
-#   r_hand_rotate,
-#   l_hand_rotate,
-#   hands_cycle
-# ]
-## Now since we're only working within a single gesture, don't need to worry about
-## Mismatched timings anymore.
-
-## see which feature is most influential in clustering
-
-## TODO limit number of clusters
-## check distance for clusters with only 1
-# https://medium.com/pythonhive/python-decorator-to-measure-the-execution-time-of-methods-fa04cb6bb36d
-
-cluster_method_timings = {'test': 0}
-
-def timeit(method):
-    def timed(*args, **kw):
-        ts = time.time()
-        result = method(*args, **kw)
-        te = time.time()
-        if 'log_time' in kw:
-            name = kw.get('log_name', method.__name__.upper())
-            kw['log_time'][name] = int((te - ts) * 1000)
-        else:
-            n = method.__name__
-            if n in cluster_method_timings.keys():
-                cluster_method_timings[n] = cluster_method_timings[n] + (te-ts) * 1000
-            else:
-                cluster_method_timings[n] = (te-ts) * 1000
-            #print('%r  %2.2f ms' % \
-            #      (method.__name__, (te - ts) * 1000))
-        return result
-    return timed
-
-
-
-
+# TODO
+# see which feature is most influential in clustering
+# add cycle detection to features
+# add rotation? arcs? sweeps? oscillation?
 
 
 ###############################################################
-####################### DIY Clustering ########################
+# DIY Clustering ##############################################
 ###############################################################
-class GestureClusterer():
+# takes vectors, normalizes across features
+def _normalize_across_features(vectors):
+    T = vectors.T
+    norms = np.array([v / np.linalg.norm(v) for v in T])
+    v_normed = norms.T
+    return v_normed
+
+
+def _create_empty_feature_vector():
+    gesture_features = [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]
+    return gesture_features
+
+
+@timeit
+def _avg(v):
+    return float(sum(v) / len(v))
+
+
+class GestureClusterer:
     # all the gesture data for gestures we want to cluster.
     # the ids of any seed gestures we want to use for our clusters.
     def __init__(self, all_gesture_data):
@@ -135,8 +64,6 @@ class GestureClusterer():
         self.cluster_file = os.path.join(homePath, "GestureData", "cluster_tmp.json")
         self.has_assigned_feature_vecs = False
         self.total_clusters_created = 0
-        ## hacky way to work around some malformatted data.
-        self.drop_ids = []
 
     def clear_clusters(self):
         self.clusters = {}
@@ -144,7 +71,9 @@ class GestureClusterer():
         self.c_id = 0
 
     @timeit
-    def cluster_gestures(self, gesture_data=None, max_cluster_distance=0.10, max_number_clusters=0, seed_ids=[]):
+    def cluster_gestures(self, gesture_data=None, max_cluster_distance=0.10, max_number_clusters=0, seed_ids=None):
+        if seed_ids is None:
+            seed_ids = []
         gd = gesture_data if gesture_data else self.agd
         if 'feature_vec' in list(gd[0].keys()):
             print("already have feature vectors in our gesture data")
@@ -159,7 +88,9 @@ class GestureClusterer():
         self._cluster_gestures(gd, max_cluster_distance, max_number_clusters, seed_ids)
 
     @timeit
-    def _cluster_gestures(self, gd, max_cluster_distance=0.03, max_number_clusters=0, seed_ids=[]):
+    def _cluster_gestures(self, gd, max_cluster_distance=0.03, max_number_clusters=0, seed_ids=None):
+        if seed_ids is None:
+            seed_ids = []
         print("Clustering gestures")
         for g in tqdm(gd):
             # if we've already seeded a cluster with this gesture, don't cluster it.
@@ -196,21 +127,22 @@ class GestureClusterer():
             self.clusters[cluster_id]['gestures'].append(g)
             self._update_cluster_centroid(cluster_id)
             g['gesture_cluster_id'] = cluster_id
-        except:
+        except RuntimeError as e:
             print('could not add gesture %s to cluster %s' % (g['id'], cluster_id))
             print('cluster keys:')
+            print(e)
             print(self.clusters[cluster_id].keys())
 
     @timeit
     def _assign_feature_vectors(self, gesture_data=None):
         gd = gesture_data if gesture_data else self.agd
-        empty_vec = self._create_empty_feature_vector()
+        empty_vec = _create_empty_feature_vector()
 
         print("Getting initial feature vectors.")
         for g in tqdm(gd):
-            if type(g['keyframes']) == type(None):
+            if isinstance(g['keyframes'], type(None)):
                 print("found empty vector")
-                ## TODO fix this
+                # TODO fix this
                 g['feature_vec'] = empty_vec
             else:
                 g['feature_vec'] = self._get_gesture_features(g)
@@ -219,8 +151,6 @@ class GestureClusterer():
         print("dropping %s empty vectors from gesture clusters" % str(len(empties)))
         # hacky ways to fix malformatted data
         self.agd = [g for g in self.agd if g['feature_vec'] != empty_vec]
-        self.agd = [g for g in self.agd if g['id'] not in self.drop_ids]
-        self.drop_ids = list(set(self.drop_ids))
         self._normalize_feature_values()
         self.has_assigned_feature_vecs = True
         return
@@ -238,18 +168,11 @@ class GestureClusterer():
         print("Normalizing feature vectors.")
         gd = gesture_data if gesture_data else self.agd
         feat_vecs = np.array([g['feature_vec'] for g in gd])
-        feat_vecs_normalized = self._normalize_across_features(feat_vecs)
+        feat_vecs_normalized = _normalize_across_features(feat_vecs)
         print("Reassigning normalized vectors")
         for i in list(range(len(gd))):
             gd[i]['feature_vec'] = list(feat_vecs_normalized[i])
         return gd
-
-    # takes vectors, normalizes across features
-    def _normalize_across_features(self, vectors):
-        T = vectors.T
-        norms = np.array([v / np.linalg.norm(v) for v in T])
-        v_normed = norms.T
-        return v_normed
 
     @timeit
     def _create_new_cluster(self, seed_gest):
@@ -268,26 +191,26 @@ class GestureClusterer():
         gd = self.agd
         print("Reclustering by centroid")
         # clear old gestures
-        randc = 0
+        rand_c = 0
         for c in self.clusters:
             self.clusters[c]['gestures'] = []
-            randc = c
+            rand_c = c
 
         for g in tqdm(gd):
             min_dist = 1000
-            min_clust = self.clusters[randc]
+            min_cluster = self.clusters[rand_c]
             for c in self.clusters:
                 d = self._calculate_distance_between_vectors(g['feature_vec'], self.clusters[c]['centroid'])
                 if d < min_dist:
                     min_dist = d
-                    min_clust = self.clusters[c]
-            min_clust['gestures'].append(g)
+                    min_cluster = self.clusters[c]
+            min_cluster['gestures'].append(g)
         return
 
     def get_sentences_by_cluster(self, cluster_id):
         c = self.clusters[cluster_id]
-        sents = [g['phase']['transcript'] for g in c['gestures']]
-        return sents
+        sentences = [g['phase']['transcript'] for g in c['gestures']]
+        return sentences
 
     def get_gesture_ids_by_cluster(self, cluster_id):
         c = self.clusters[cluster_id]
@@ -297,7 +220,7 @@ class GestureClusterer():
         #     ids.append(g['id'])
         return ids
 
-    def report_clusters(self, verbose=False):
+    def report_clusters(self):
         print(("Number of clusters: %s" % len(self.clusters)))
         cluster_rep = [(c, len(self.clusters[c]['gestures'])) for c in list(self.clusters.keys())]
         cluster_lengths = [len(self.clusters[c]['gestures']) for c in list(self.clusters.keys())]
@@ -306,7 +229,7 @@ class GestureClusterer():
         print(("Median cluster size: %s" % np.median(cluster_lengths)))
         print(("Largest cluster size: %s" % max(cluster_lengths)))
         cluster_sparsity = [self.get_cluster_sparsity(c) for c in list(self.clusters.keys())]
-        print(("Cluster sparsities: %s" % cluster_sparsity))
+        print(("Cluster sparsity: %s" % cluster_sparsity))
         print(("Avg cluster sparsity: %s" % np.average(cluster_sparsity)))
         print(("Median cluster sparsity: %s" % np.median(cluster_sparsity)))
         print(("Sanity check: total clustered gestures: %s / %s" % (sum(cluster_lengths), len(self.agd))))
@@ -315,16 +238,14 @@ class GestureClusterer():
         # TODO: also get minimum and maximum centroid distances.
         return self.clusters
 
-
-    ## measure of how distant the gestures are... basically avg distance to centroid
+    # measure of how distant the gestures are... basically avg distance to centroid
     def get_cluster_sparsity(self, cluster_id):
         c = self.clusters[cluster_id]
         cent = c['centroid']
         dists = [self._calculate_distance_between_vectors(g['feature_vec'], cent) for g in c['gestures']]
         return np.average(dists)
 
-
-    ## instead of this need to use centroid.
+    # instead of this need to use centroid.
     @timeit
     def _get_shortest_cluster_dist(self, g):
         shortest_dist = 10000
@@ -336,14 +257,14 @@ class GestureClusterer():
             if dist < shortest_dist:
                 nearest_cluster_id = c['cluster_id']
             shortest_dist = min(shortest_dist, dist)
-        return (nearest_cluster_id, shortest_dist)
+        return nearest_cluster_id, shortest_dist
 
-    ## TODO check this bad boy for bugs.
+    # TODO check this bad boy for bugs.
     def _update_cluster_centroid(self, cluster_id):
         s = time.time()
         c = self.clusters[cluster_id]
-        ## very very slow.
-        ## TODO speed this up using matrix magic.
+        # very very slow.
+        # TODO speed this up using matrix magic.
         feat_vecs = [g['feature_vec'] for g in c['gestures']]
         feat_vecs = np.array(feat_vecs[0])
         self._log("old centroid: %s" % c['centroid'])
@@ -351,29 +272,29 @@ class GestureClusterer():
         self._log("new centroid: %s" % c['centroid'])
         self.clusters[cluster_id] = c
         e = time.time()
-        self._log("time to update centroid: %s" % str(e-s))
+        self._log("time to update centroid: %s" % str(e - s))
         return
 
     ############################################################
-    #################### MOVEMENT CHARACTERISTICS ##############
+    # MOVEMENT CHARACTERISTICS #################################
     ############################################################
     @timeit
     def _get_gesture_features(self, gesture, gesture_features=GESTURE_FEATURES):
         r_keyframes = self._get_rl_hand_keypoints(gesture, 'r')
         l_keyframes = self._get_rl_hand_keypoints(gesture, 'l')
-        gesture_features = []
+        feature_vector = []
         for feature in gesture_features:
             if not GESTURE_FEATURES[feature]['separate_hands']:
                 val = GESTURE_FEATURES[feature]['function'](r_keyframes, l_keyframes)
-                gesture_features.append(val)
+                feature_vector.append(val)
             else:
                 val1 = GESTURE_FEATURES[feature]['function'](r_keyframes)
                 val2 = GESTURE_FEATURES[feature]['function'](l_keyframes)
-                gesture_features.append(val1)
-                gesture_features.append(val2)
-        return gesture_features
+                feature_vector.append(val1)
+                feature_vector.append(val2)
+        return feature_vector
 
-        #gesture_features = [
+        # gesture_features = [
         #  self._palm_vert(l_keyframes),
         #  self._palm_horiz(l_keyframes),
         #  self._palm_vert(r_keyframes),
@@ -395,12 +316,8 @@ class GestureClusterer():
         #   r_hand_rotate,
         #   l_hand_rotate,
         #   hands_cycle
-        #]
-        #return gesture_features
-
-    def _create_empty_feature_vector(self):
-        gesture_features = [0,0,0,0,0,0,0,0,0,0,0,0]
-        return gesture_features
+        # ]
+        # return gesture_features
 
     # TODO manually make some features more/less important
     # report importance of each individual feature in clustering
@@ -408,25 +325,21 @@ class GestureClusterer():
 
     # TODO check what audio features are in original paper
 
-    ## Audio for agent is bad -- TTS is garbaggio
-    ## assumes there is good prosody in voice (TTS there isn't)
+    # Audio for agent is bad -- TTS is garbaggio
+    # assumes there is good prosody in voice (TTS there isn't)
 
-    ## Co-optimize gesture-language clustering (learn how)
-    ## KL distance for two clusters?
+    # Co-optimize gesture-language clustering (learn how)
+    # KL distance for two clusters?
 
     # learn similarity of sentences from within one gesture
     # how to map gesture clusters <--> sentence clusters
-    ## in the end want to optimize overlapping clusters btw gesture/language
+    # in the end want to optimize overlapping clusters btw gesture/language
 
     # probabilistic mapping of sentence (from gesture) to sentence cluster
 
     ##############################################################
-    ####################### Helpers/Calcs ########################
+    # Helpers/Calculators ########################################
     ##############################################################
-    @timeit
-    def _avg(self, v):
-        return float(sum(v) / len(v))
-
     @timeit
     def _get_rl_hand_keypoints(self, gesture, hand):
         keys = []
@@ -440,14 +353,7 @@ class GestureClusterer():
             if type(t) != dict:
                 print("found empty keyframes for gesture %s" % gesture['id'])
                 print(t)
-                print(gesture)
-                self.drop_ids.append(gesture['id'])
-                t = {
-                    'x': [],
-                    'y': []
-                }
-                ## KNOWN TEMP FIX
-                return [{'y':[0], 'x':[0]}]
+                return [{'y': [0], 'x': [0]}]
             y = [t['y'][i] for i in keypoint_range]
             x = [t['x'][i] for i in keypoint_range]
             keys.append({'y': y, 'x': x})
@@ -460,7 +366,7 @@ class GestureClusterer():
 
         feat1 = np.array(self._get_gesture_features(g1))
         feat2 = np.array(self._get_gesture_features(g2))
-        return np.linalg.norm(feat1-feat2)
+        return np.linalg.norm(feat1 - feat2)
 
     @timeit
     def _calculate_distance_between_vectors(self, v1, v2):
@@ -471,8 +377,8 @@ class GestureClusterer():
 
     def _write_logs(self):
         with open(self.logfile, 'w') as f:
-            for l in self.logs:
-                f.write("%s\n" % l)
+            for log in self.logs:
+                f.write("%s\n" % log)
         f.close()
 
     def get_closest_gesture_to_centroid(self, cluster_id):
@@ -484,7 +390,7 @@ class GestureClusterer():
             dist = self._calculate_distance_between_vectors(g['feature_vec'], cent)
             if dist < min_d:
                 g_id = g['id']
-                dist = min_d
+                min_d = dist
         return g_id
 
     def get_random_gesture_id_from_cluster(self, cluster_id):
@@ -520,12 +426,11 @@ class GestureClusterer():
             for j in range(len(gs)):
                 if i == j:
                     continue
-                all_dists = all_dists + self._calculate_distance_between_vectors(gs[i]['feature_vec'], gs[j]['feature_vec'])
+                all_dists = all_dists + self._calculate_distance_between_vectors(gs[i]['feature_vec'],
+                                                                                 gs[j]['feature_vec'])
 
-        dists = {}
-        dists['average'] = self._avg(all_dists)
-        dists['max'] = max(all_dists)
-        dists['min'] = min(all_dists)
+        dists = {'average': _avg(all_dists), 'max': max(all_dists), 'min': min(all_dists)}
+        return dists
 
     # takes a cluster ID, returns nearest neighbor cluster ID
     def get_nearest_cluster_id(self, cluster_id):
@@ -534,23 +439,12 @@ class GestureClusterer():
         for c in self.clusters:
             if c == cluster_id:
                 continue
-            mind = self._calculate_distance_between_vectors(self.clusters[c]['centroid'], self.clusters[cluster_id]['centroid'])
+            mind = self._calculate_distance_between_vectors(self.clusters[c]['centroid'],
+                                                            self.clusters[cluster_id]['centroid'])
             if mind < dist:
                 dist = mind
                 min_c = c
         return min_c
-
-    # takes cluster ID, returns nearest neighbor cluster distance
-    def get_nearest_cluster_distance(self, cluster_id):
-        dist = 1000
-        for c in self.clusters:
-            if c == cluster_id:
-                continue
-            mind = self._calculate_distance_between_vectors(self.clusters[c]['centroid'], self.clusters[cluster_id]['centroid'])
-            if mind < dist:
-                dist = mind
-                min_c = c
-        return dist
 
     # given a point and cluster id, returns avg distance between the point and
     # all points in the cluster.
@@ -563,22 +457,21 @@ class GestureClusterer():
             return 0
         for g in c['gestures']:
             dists.append(self._calculate_distance_between_vectors(vec, g['feature_vec']))
-        return self._avg(dists)
+        return _avg(dists)
 
     # gets silhouette score for cluster using centroid
     def get_silhouette_score(self, cluster_id):
         p = self.clusters[cluster_id]['centroid']
         a = self.get_avg_dist_between_point_and_cluster(p, cluster_id)
         b = self.get_avg_dist_between_point_and_cluster(p, self.get_nearest_cluster_id(cluster_id))
-        score = (b - a)/max(b,a)
+        score = (b - a) / max(b, a)
         return score
 
     def get_avg_silhouette_score(self):
         scores = []
         for g in self.clusters:
             scores.append(self.get_silhouette_score(g))
-        return self._avg(scores)
-
+        return _avg(scores)
 
 # our basic problem is that we need to figure out how to map distances between motions that are very long
 # vectors, and different lengths of keyframes. But we need to distinguish between the speed of those motions
@@ -590,269 +483,3 @@ class GestureClusterer():
 
 
 # Silhouette scores for clusters are a good way of determining how many "base" gestures there may be??
-
-
-##########################################################################
-####################### STATIC MOTION FUNCTIONS ##########################
-##########################################################################
-GESTURE_FEATURES = {
-    'palm_vertical': {
-        'separate_hands': True,
-        'function': _palm_vert
-    },
-    'palm_horizontal': {
-        'separate_hands': True,
-        'function': _palm_horiz
-    },
-    'max_hands_apart': {
-        'separate_hands': False,
-        'function': _max_hands_apart
-    },
-    'min_hands_together': {
-        'separate_hands': False,
-        'function': _min_hands_together
-    },
-    'wrists_up': {
-        'separate_hands': True,
-        'function': _wrists_up
-    },
-    'wrists_down': {
-        'separate_hands': True,
-        'function': _wrists_down
-    },
-    'wrists_apart': {
-        'separate_hands': False,
-        'function': _wrists_apart
-    },
-    'wrists_together': {
-        'separate_hands': False,
-        'function': _wrists_together
-    },
-    'max_wrist_velocity': {
-        'separate_hands': True,
-        'function': _max_wrist_velocity
-    }
-    # 'acceleration': {
-    #     'separate_hands': True,
-    #     'function': _max_wrist_velocity
-    # }
-    # 'oscillation': {
-    #     'separate_hands': True,
-    #     'function': _max_wrist_velocity
-    # }
-    # 'hand_position_changes': {
-    #     'separate_hands': True,
-    #     'function': _max_wrist_velocity
-    # }
-}
-
-
-# get maximum "verticalness" aka minimum horizontalness of hands
-def _palm_vert(keyframes):
-    return _palm_angle_axis(keyframes, 'x')
-
-
-# get maximum "horizontalness" aka minimum verticalness of hands
-def _palm_horiz(keyframes):
-    return _palm_angle_axis(keyframes, 'y')
-
-
-@timeit
-def _palm_angle_axis(keyframes, xy):
-    p_min = 1000
-    for frame in keyframes:
-        max_frame_dist = max(frame[xy]) - min(frame[xy])
-        p_min = min(p_min, max_frame_dist)
-    return p_min
-
-
-## max distance from low --> high the wrists move in a single stroke
-def _wrists_up(keyframes):
-    return _wrist_vertical_stroke(keyframes, operator.ge)
-
-
-## max distance from high --> low the wrists move in single stroke
-def _wrists_down(keyframes):
-    return _wrist_vertical_stroke(keyframes, operator.le)
-
-
-@timeit
-def _wrist_vertical_stroke(keyframes, relate):
-    total_motion = 0
-    max_single_stroke = 0
-    # 0 is the index of the wrist in handed keypoints
-    pos = keyframes[0]['y'][0]
-    same_direction = False
-    for frame in keyframes:
-        curr_pos = frame['y'][0]
-        if relate(curr_pos, pos):
-            total_motion = total_motion + abs(curr_pos - pos)
-            pos = curr_pos
-            same_direction = True
-        else:
-            if same_direction:
-                total_motion = 0
-            same_direction = False
-            pos = curr_pos
-        max_single_stroke = max(max_single_stroke, total_motion)
-    return max_single_stroke
-
-
-# returns minimum distance at any frame between point A on right hand and
-# point A on left hand.
-def _min_hands_together(r_hand_keys, l_hand_keys):
-    return _max_hand_togetherness(r_hand_keys, l_hand_keys, 10000, min)
-
-
-# returns maximum distance at any frame between point A on right hand and
-# point A on left hand
-def _max_hands_apart(r_hand_keys, l_hand_keys):
-    return _max_hand_togetherness(r_hand_keys, l_hand_keys, 0, max)
-
-
-@timeit
-def _max_hand_togetherness(r_hand_keys, l_hand_keys, min_max, relate):
-    max_dist = min_max          # larger than pixel range
-    for frame_index in range(len(r_hand_keys)):     # for each frame in the gesture, better be same cause it's the same gesture!
-        # determine how far apart r and l hand are, vertically and horizontally
-        cur_r_keys = r_hand_keys[frame_index]
-        cur_l_keys = l_hand_keys[frame_index]
-        r_pos = np.array((cur_r_keys['x'], cur_r_keys['y']))
-        l_pos = np.array((cur_l_keys['x'], cur_l_keys['y']))
-        dist = np.linalg.norm(r_pos-l_pos)
-        max_dist = relate(dist, max_dist)
-    return max_dist
-
-# measures the largest outward motion of r/l wrists
-# that is, the largest distance in which wrists are moving
-# continuously apart.
-# THIS IS IN SPACE, USES BOTH HORIZ AND VERT AXES
-def _wrists_apart(r_hand_keys, l_hand_keys):
-    return _wrist_togetherness(r_hand_keys, l_hand_keys, operator.ge)
-
-def _wrists_together(r_hand_keys, l_hand_keys):
-    return _wrist_togetherness(r_hand_keys, l_hand_keys, operator.le)
-
-@timeit
-def _wrist_togetherness(r_hand_keys, l_hand_keys, relate):
-    total_direction_dist = 0
-    max_direction_dist = 0
-    # the 0th keyframe of each hand is the wrist position
-    r_wrist_position = np.array([r_hand_keys[0]['x'][0], r_hand_keys[0]['y'][0]])
-    l_wrist_position = np.array([l_hand_keys[0]['x'][0], l_hand_keys[0]['y'][0]])
-    prev_dist = np.linalg.norm(r_wrist_position - l_wrist_position)
-    for frame_index in range(len(r_hand_keys)):
-        r_wrist_position = np.array([r_hand_keys[frame_index]['x'][0], r_hand_keys[frame_index]['y'][0]])
-        l_wrist_position = np.array([l_hand_keys[frame_index]['x'][0], l_hand_keys[frame_index]['y'][0]])
-        cur_dist = np.linalg.norm(r_wrist_position - l_wrist_position)
-        if relate(cur_dist, prev_dist):     # we are moving in the desired direction
-            total_direction_dist += abs(cur_dist - prev_dist)
-            prev_dist = cur_dist
-            max_direction_dist = max(max_direction_dist, total_direction_dist)
-        else:                       # we're not moving in the desired direction
-            total_direction_dist = 0
-            prev_dist = cur_dist
-    return max_direction_dist
-
-
-# max velocity of wrist between 2 frames
-# specifically, it just gets the max difference in distance between wrists across 2 frames
-@timeit
-def _max_wrist_velocity(keys):
-    max_dist = 0
-    for i in range(len(keys)-1):
-        # wrist is 0th keypoint for each hand
-        (wx0, wy0) = (keys[i]['x'][0], keys[i]['y'][0])
-        (wx1, wy1) = (keys[i+1]['x'][0], keys[i+1]['y'][0])
-        max_dist = max(max_dist, _get_point_dist(wx0, wy0, wx1, wy1))
-    return max_dist
-
-
-# measures the number of times wrist changes direction as measured by the angle
-# of the wrist point between frame a, b, c is greater than 100
-def _get_back_and_forth(keys):
-    switches = 0
-    if len(keys) < 3:
-        return 0
-    for frame in range(len(keys)-2):
-        a = np.array([keys[frame]['x'][0], keys[frame]['y'][0]])
-        b = np.array([keys[frame+1]['x'][0], keys[frame+1]['y'][0]])
-        c = np.array([keys[frame+2]['x'][0], keys[frame+2]['y'][0]])
-        ba = a - b
-        bc = c - b
-        cos_ang = np.dot(ba, bc) / (np.linalg.norm(ba) * np.linalg.norm(bc))
-        full_ang = np.arccos(cos_ang)
-        if np.degrees(full_ang) >= DIRECTION_ANGLE_SWITCH:      # arbitrary measure for switching directions ¯\_(ツ)_/¯
-            switches += 1
-    return switches
-
-
-# TODO check this in MotionAnalyzerTests and visually.
-# actually right now seems to detect smooth gestures???
-# definitely not right.
-def _max_acceleration(keys):
-    max_accel = 0
-    for frame in range(len(keys)-2):
-        ax, ay = keys[frame]['x'][0], keys[frame]['y'][0]
-        bx, by = keys[frame+1]['x'][0], keys[frame+1]['y'][0]
-        cx, cy = keys[frame+2]['x'][0], keys[frame+2]['y'][0]
-        d1 = _get_point_dist(ax, ay, bx, by)   # /1 for 1 frame (velocity, not distance technically)
-        d2 = _get_point_dist(bx, by, cx, cy)
-        max_accel = max(max_accel, abs(d1-d2))
-    return max_accel
-
-
-# given a set of keys from a hand (array length 22), returns angles between every 3 points, like trigrams
-# works on frame i
-# if hand angles are roughly the same (within like, 20 degrees for each thing) then they're about the same shape
-    # TODO map angles against frame...
-def _get_hand_angles_for_frame(handed_keys, frame_index):
-    # calculate angles for
-    # 0,1,2,3,4
-    # 0,5,6,7,8
-    # 0,9,10,11,12
-    # 0,13,14,15,16
-    # 0,17,18,19,20
-    # for each of these calc between 0-1-2, 1-2-3, 2-3-4
-    # TODO clean this up
-    angles = []
-    kf = handed_keys[frame_index]
-    for i in range(5):      # 5 fingers
-        # for j in range(3):      # each of the angles on the fingers (0,1,2; 1,2,3; 2,3,4)
-        #    a = np.array((kf['x'][(i * 4) + j], kf['y'][(i * 4) + j]))
-        #    b = np.array((kf['x'][(i * 4) + j+1], kf['y'][(i * 4) + j+1]))
-        #    c = np.array((kf['x'][(i * 4) + j+2], kf['y'][(i * 4) + j+2]))
-        #    ab = a - b
-        #    cb = c - b
-        #    cos_b = np.dot(ab, cb) / (np.linalg.norm(ab) * np.linalg.norm(cb))
-        #    ang_b = np.arccos(cos_b)
-        #    angles.append(np.degrees(ang_b))
-        # return angles
-        base = np.array((kf['x'][0], kf['y'][0]))
-        a = np.array((kf['x'][(i*4)+1], kf['y'][(i*4)+1]))
-        b = np.array((kf['x'][(i*4)+2], kf['y'][(i*4)+2]))
-        c = np.array((kf['x'][(i*4)+3], kf['y'][(i*4)+3]))
-        d = np.array((kf['x'][(i*4)+4], kf['y'][(i*4)+4]))
-        basea = base - a
-        ba = b - a
-        cos_a = np.dot(basea, ba) / (np.linalg.norm(basea) * np.linalg.norm(ba))
-        ang_a = np.arccos(cos_a)
-        angles.append(np.degrees(ang_a))
-        ab = a - b
-        cb = c - b
-        cos_b = np.dot(ab, cb) / (np.linalg.norm(ab) * np.linalg.norm(cb))
-        ang_b = np.arccos(cos_b)
-        angles.append(np.degrees(ang_b))
-        bc = b - c
-        dc = d - c
-        cos_c = np.dot(bc, dc) / (np.linalg.norm(bc) * np.linalg.norm(dc))
-        ang_c = np.arccos(cos_c)
-        angles.append(np.degrees(ang_c))
-    return angles
-
-
-@timeit
-def _get_point_dist(x1,y1,x2,y2):
-    a = np.array((x1, y1))
-    b = np.array((x2, y2))
-    return np.linalg.norm(a-b)
